pyproject.toml
[tool.setuptools]
packages = ["aquaspot"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aquaspot"
version = "0.1.0"
description = "Pipeline leak detection using Sentinel-2 NDWI analysis"
authors = [{name = "AquaSpot Team", email = "team@aquaspot.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rasterio>=1.3.0",
    "shapely>=2.0.0",
    "geopandas>=0.14.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "scikit-image>=0.21.0",
    "streamlit>=1.28.0",
    "weasyprint>=60.0",
    "jinja2>=3.1.0",
    "matplotlib>=3.7.0",
    "folium>=0.14.0",
    "pystac-client>=0.7.0",
    "planetary-computer>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "requests-mock>=1.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
alerts = [
    "twilio>=8.0.0",
]

[project.scripts]
aquaspot = "aquaspot.cli:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["S101", "S602", "S603", "G004", "PLR2004", "ARG001", "ERA001", "FBT001", "E501", "PTH123", "DTZ001", "NPY002", "INP001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
